+++ packages/atom.io/src/internal/index.ts
@@ -40,11 +40,11 @@ export * from "./join"
 export * from "./junction"
 export * from "./keys"
 export * from "./lineage"
-export * from "./map-overlay"
 export * from "./molecule"
 export * from "./mutable"
 export * from "./not-found-error"
 export * from "./operation"
+export * from "./overlays"
 export * from "./reserved-keys"
 export * from "./selector"
 export * from "./set-state"
