+++ packages/atom.io/src/internal/overlays/index.ts
@@ -1,0 +1,3 @@
+export * from "./map-overlay"
+export * from "./relations-overlay"
+export * from "./set-overlay"
