+++ packages/atom.io/src/internal/transaction/build-transaction.ts
@@ -12,8 +12,8 @@ import { disposeFromStore, findInStore } from "../families"
 import { getEnvironmentData } from "../get-environment-data"
 import { getFromStore } from "../get-state"
 import { newest } from "../lineage"
-import { MapOverlay } from "../map-overlay"
 import { getJsonToken } from "../mutable"
+import { MapOverlay } from "../overlays/map-overlay"
 import { resetInStore, setIntoStore } from "../set-state"
 import type { Store } from "../store"
 import type { Fn } from "../utility-types"
